+++ apps/admin-ui/src/App.tsx
@@ -1,20 +1,18 @@
 import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import dynamic from "next/dynamic";
-
-import { Permission } from "app/modules/permissionHelper";
-import ApplicationRound from "./spa/recurring-reservations/application-rounds/[id]";
-import PageWrapper from "./component/PageWrapper";
+import { Permission } from "@modules/permissionHelper";
+import PageWrapper from "@component/PageWrapper";
 import "./i18n";
-import { publicUrl } from "./common/const";
-import { GlobalContext } from "./context/GlobalContexts";
-import { prefixes } from "./common/urls";
-import AuthorizationChecker from "./common/AuthorizationChecker";
-
-import MyUnitsRouter from "./component/my-units/MyUnitsRouter";
-import ReservationsRouter from "./component/reservations/ReservationRouter";
-import NotificationsRouter from "./component/notifications/router";
-import Error404 from "./common/Error404";
+import { publicUrl } from "@common/const";
+import { GlobalContext } from "@context/GlobalContexts";
+import { prefixes } from "@common/urls";
+import { AuthorisationChecker } from "@component/AuthorisationChecker";
+import MyUnitsRouter from "@component/my-units/MyUnitsRouter";
+import ReservationsRouter from "@component/reservations/ReservationRouter";
+import NotificationsRouter from "@component/notifications/router";
+import { Error404 } from "@component/error";
+import ApplicationRound from "./spa/recurring-reservations/application-rounds/[id]";
 
 const UNIT_PATH = "./component/Unit";
 const Units = dynamic(() => import(`${UNIT_PATH}/Units`));
@@ -58,9 +56,9 @@ const ApplicationRoundAllocation = dynamic(
 );
 
 const withAuthorization = (component: JSX.Element, permission?: Permission) => (
-  <AuthorizationChecker permission={permission}>
+  <AuthorisationChecker permission={permission}>
     {component}
-  </AuthorizationChecker>
+  </AuthorisationChecker>
 );
 
 const UnitsRouter = () => (
