--- apps/admin-ui/src/common/hds-fork/Accordion.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { Accordion as HDSAccordion, AccordionProps } from "hds-react";
-
-const Accordion = ({ ...props }: AccordionProps) => (
-  <HDSAccordion closeButton={false} {...props} />
-);
-
-export { Accordion };
