+++ apps/admin-ui/src/component/PageWrapper.tsx
@@ -2,13 +2,12 @@ import React, { Suspense } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 import * as Sentry from "@sentry/nextjs";
 import styled from "styled-components";
-
 import ClientOnly from "common/src/ClientOnly";
-import Error5xx from "app/common/Error5xx";
-import usePermission from "app/hooks/usePermission";
 import { BannerNotificationsList } from "common/src/components";
 import { CommonBannerNotificationTargetChoices } from "common/types/gql-types";
-import ScrollToTop from "../common/ScrollToTop";
+import { ScrollToTop } from "@/component/ScrollToTop";
+import { Error5xx } from "@/component/error";
+import usePermission from "@/hooks/usePermission";
 import GlobalElements from "./GlobalElements";
 import Navigation from "./Navigation";
 import MainMenu from "./MainMenu";
