+++ apps/admin-ui/src/component/ScrollIntoView.tsx
@@ -5,7 +5,9 @@ interface Props {
   children: JSX.Element;
 }
 
-const ScrollIntoView = ({ hash, children }: Props): JSX.Element => {
+// TODO refactor to a hook
+// TODO there is a duplicate in the ui app
+export function ScrollIntoView({ hash, children }: Props): JSX.Element {
   const selfRef = React.useRef<HTMLDivElement | null>(null);
 
   const isMatch = hash === document.location.hash?.substring(1);
@@ -24,6 +26,4 @@ const ScrollIntoView = ({ hash, children }: Props): JSX.Element => {
   ) : (
     children
   );
-};
-
-export default ScrollIntoView;
+}
