+++ apps/admin-ui/src/component/error/index.ts
@@ -1,0 +1,3 @@
+export * from "./Error403";
+export * from "./Error404";
+export * from "./Error5xx";
