+++ apps/admin-ui/src/pages/_app.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import type { AppProps } from "next/app";
 import "hds-core/lib/base.css";
 import "../index.scss";
-import { isBrowser } from "app/common/const";
+import { isBrowser } from "@common/const";
 import { ApolloProvider } from "@apollo/client";
-import apolloClient from "app/common/apolloClient";
-import ExternalScripts from "../common/ExternalScripts";
+import apolloClient from "@common/apolloClient";
+import { ExternalScripts } from "@component/ExternalScripts";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
