+++ apps/admin-ui/src/spa/recurring-reservations/application-rounds/index.tsx
@@ -12,7 +12,7 @@ import {
 import { filterNonNullable } from "common/src/helpers";
 import { applicationRoundUrl } from "@/common/urls";
 import { formatDate } from "@/common/util";
-import { Accordion } from "@/common/hds-fork/Accordion";
+import { Accordion } from "hds-react";
 import { useNotification } from "@/context/NotificationContext";
 import { Container } from "@/styles/layout";
 import { truncate } from "@/helpers";
@@ -56,7 +56,7 @@ const RoundsAccordion = ({
   }
 
   return (
-    <Accordion heading={name} initiallyOpen={initiallyOpen}>
+    <Accordion heading={name} initiallyOpen={initiallyOpen} closeButton={false}>
       <AccordionContainer>
         {!rounds || rounds.length === 0
           ? emptyContent || <span>no data {name}</span>
